test(web): split template render suite into focused modules [OPE-386]
