--- crates/opengoose-web/src/tests/template_render.rs
@@ -1,535 +1,0 @@
-use crate::data::{
-    BatchExportFormView, MessageBubble, MetaRow, MetricCard, Notice, PluginDetailView,
-    PluginFilterItem, PluginListItem, PluginsPageView, QueueDetailView, QueueMessageView,
-    ScheduleEditorView, ScheduleHistoryItem, ScheduleListItem, SchedulesPageView, SelectOption,
-    SessionDetailView, SessionExportAction, SessionListItem, SessionsPageView, TriggerDetailView,
-    TriggerListItem, TriggersPageView, WorkflowAutomationView, WorkflowDetailView,
-    WorkflowListItem, WorkflowRunView, WorkflowStepView, WorkflowsPageView,
-};
-use crate::fixtures::sample_dashboard_view;
-use crate::routes;
-
-fn sample_session_detail() -> SessionDetailView {
-    SessionDetailView {
-        session_key: "discord:ops:bridge".into(),
-        title: "Session ops".into(),
-        subtitle: "discord / ops".into(),
-        source_label: "Live runtime".into(),
-        export_actions: vec![
-            SessionExportAction {
-                label: "Export JSON".into(),
-                href: "/api/sessions/discord%3Aops%3Abridge/export?format=json".into(),
-            },
-            SessionExportAction {
-                label: "Export Markdown".into(),
-                href: "/api/sessions/discord%3Aops%3Abridge/export?format=md".into(),
-            },
-        ],
-        meta: vec![MetaRow {
-            label: "Stable key".into(),
-            value: "discord:ops:bridge".into(),
-        }],
-        notice: None,
-        selected_model: String::new(),
-        model_options: vec![],
-        messages: vec![MessageBubble {
-            role_label: "Assistant".into(),
-            author_label: "frontend-engineer".into(),
-            timestamp: "2026-03-10 12:34".into(),
-            content: "Dashboard panel updated.".into(),
-            tone: "accent",
-            alignment: "right",
-        }],
-        empty_hint: "No messages yet.".into(),
-    }
-}
-
-fn sample_queue_detail() -> QueueDetailView {
-    QueueDetailView {
-        title: "Queue run-1".into(),
-        subtitle: "feature-dev / chain".into(),
-        source_label: "Live runtime".into(),
-        status_cards: vec![MetricCard {
-            label: "Pending".into(),
-            value: "1".into(),
-            note: "Waiting for recipients".into(),
-            tone: "amber",
-        }],
-        messages: vec![QueueMessageView {
-            sender: "planner".into(),
-            recipient: "developer".into(),
-            kind: "task".into(),
-            status_label: "pending".into(),
-            status_tone: "amber",
-            created_at: "2026-03-10 12:35".into(),
-            retry_text: "1/3".into(),
-            content: "Implement the dashboard controls.".into(),
-            error: "waiting for pickup".into(),
-        }],
-        dead_letters: vec![],
-        empty_hint: "No queue traffic yet.".into(),
-    }
-}
-
-fn sample_plugin_detail() -> PluginDetailView {
-    PluginDetailView {
-        title: "ops-tools".into(),
-        subtitle: "Operational helpers for dashboards and workflows.".into(),
-        source_label: "/home/dh/.opengoose/plugins/ops-tools".into(),
-        status_label: "Ready".into(),
-        status_tone: "success",
-        lifecycle_label: "Enabled".into(),
-        lifecycle_tone: "sage",
-        runtime_label: "Runtime initialized".into(),
-        runtime_tone: "success",
-        status_summary: "1 declared skill is registered in the active runtime.".into(),
-        runtime_note: Some("registered 1 declared skill(s)".into()),
-        meta: vec![
-            MetaRow {
-                label: "Version".into(),
-                value: "1.2.3".into(),
-            },
-            MetaRow {
-                label: "Runtime".into(),
-                value: "Runtime initialized".into(),
-            },
-        ],
-        capabilities: vec!["skill".into(), "channel_adapter".into()],
-        capabilities_hint: "No capabilities declared.".into(),
-        registered_skills: vec!["ops-tools/check-health".into()],
-        missing_skills: vec![],
-        notice: Some(Notice {
-            text: "Installed plugin `ops-tools`.".into(),
-            tone: "success",
-        }),
-        install_source_path: String::new(),
-        toggle_label: "Disable plugin".into(),
-        delete_label: "ops-tools".into(),
-        is_placeholder: false,
-    }
-}
-
-fn sample_workflow_detail() -> WorkflowDetailView {
-    WorkflowDetailView {
-        title: "feature-dev".into(),
-        subtitle: "Build and verify product changes with a chained team.".into(),
-        source_label: "Live registry".into(),
-        status_label: "Running".into(),
-        status_tone: "cyan",
-        meta: vec![MetaRow {
-            label: "Pattern".into(),
-            value: "Chain".into(),
-        }],
-        steps: vec![WorkflowStepView {
-            title: "Step 1 · planner".into(),
-            detail: "Shape the implementation plan.".into(),
-            badge: "Sequential".into(),
-            badge_tone: "cyan",
-        }],
-        automations: vec![WorkflowAutomationView {
-            kind: "Schedule".into(),
-            title: "nightly-review".into(),
-            detail: "0 0 * * * · team feature-dev".into(),
-            note: "Next 2026-03-11 00:00".into(),
-            status_label: "Enabled".into(),
-            status_tone: "sage",
-        }],
-        recent_runs: vec![WorkflowRunView {
-            title: "Run run-1".into(),
-            detail: "2/4 steps · Still executing".into(),
-            updated_at: "2026-03-10 12:35".into(),
-            status_label: "Running".into(),
-            status_tone: "cyan",
-            page_url: "/runs?run=run-1".into(),
-        }],
-        yaml: "title: feature-dev".into(),
-        trigger_api_url: "/api/workflows/feature-dev/trigger".into(),
-        trigger_input: "Manual run requested".into(),
-    }
-}
-
-fn sample_schedule_detail() -> ScheduleEditorView {
-    ScheduleEditorView {
-        title: "nightly-review".into(),
-        subtitle: "Adjust cadence, target team, and run input without leaving the dashboard."
-            .into(),
-        source_label: "Live schedule store".into(),
-        original_name: "nightly-review".into(),
-        name: "nightly-review".into(),
-        cron_expression: "0 0 * * * *".into(),
-        team_name: "feature-dev".into(),
-        input: String::new(),
-        enabled: true,
-        is_new: false,
-        name_locked: true,
-        meta: vec![MetaRow {
-            label: "Next fire".into(),
-            value: "2026-03-11 00:00:00".into(),
-        }],
-        team_options: vec![SelectOption {
-            value: "feature-dev".into(),
-            label: "feature-dev".into(),
-            selected: true,
-        }],
-        history: vec![ScheduleHistoryItem {
-            title: "run-1".into(),
-            detail: "chain workflow · Scheduled run: nightly-review".into(),
-            updated_at: "2026-03-10 12:35".into(),
-            status_label: "completed".into(),
-            status_tone: "sage",
-            page_url: "/runs?run=run-1".into(),
-        }],
-        history_hint: "No matching runs found for this schedule yet.".into(),
-        notice: Some(Notice {
-            text: "Schedule saved.".into(),
-            tone: "success",
-        }),
-        save_label: "Save changes".into(),
-        toggle_label: "Pause schedule".into(),
-        delete_label: "nightly-review".into(),
-    }
-}
-
-fn sample_new_schedule_detail() -> ScheduleEditorView {
-    ScheduleEditorView {
-        title: "New schedule".into(),
-        subtitle: "Create a cron-driven workflow handoff from the dashboard.".into(),
-        source_label: "Live schedule store".into(),
-        original_name: "__new__".into(),
-        name: String::new(),
-        cron_expression: String::new(),
-        team_name: String::new(),
-        input: "{}".into(),
-        enabled: false,
-        is_new: true,
-        name_locked: false,
-        meta: vec![MetaRow {
-            label: "Next fire".into(),
-            value: "Pending cron validation".into(),
-        }],
-        team_options: vec![],
-        history: vec![],
-        history_hint: "No matching runs found for this schedule yet.".into(),
-        notice: None,
-        save_label: "Create schedule".into(),
-        toggle_label: "Enable schedule".into(),
-        delete_label: "New schedule".into(),
-    }
-}
-
-fn sample_trigger_detail() -> TriggerDetailView {
-    TriggerDetailView {
-        name: "on-pr-open".into(),
-        trigger_type: "webhook_received".into(),
-        team_name: "code-review".into(),
-        input: "Review the latest pull request".into(),
-        condition_json: r#"{"path":"/pr"}"#.into(),
-        enabled: true,
-        fire_count: 4,
-        last_fired_at: "2026-03-10 12:34".into(),
-        created_at: "2026-03-09 09:00".into(),
-        meta: vec![
-            MetaRow {
-                label: "Type".into(),
-                value: "Webhook".into(),
-            },
-            MetaRow {
-                label: "Team".into(),
-                value: "code-review".into(),
-            },
-        ],
-        status_label: "enabled".into(),
-        status_tone: "success",
-        notice: Some(Notice {
-            text: "Trigger saved.".into(),
-            tone: "success",
-        }),
-        is_placeholder: false,
-    }
-}
-
-fn sample_trigger_placeholder_detail() -> TriggerDetailView {
-    TriggerDetailView {
-        name: String::new(),
-        trigger_type: String::new(),
-        team_name: String::new(),
-        input: String::new(),
-        condition_json: "{}".into(),
-        enabled: false,
-        fire_count: 0,
-        last_fired_at: "never".into(),
-        created_at: String::new(),
-        meta: vec![],
-        status_label: "none".into(),
-        status_tone: "neutral",
-        notice: Some(Notice {
-            text: "Name, type, and team are required to create a trigger.".into(),
-            tone: "danger",
-        }),
-        is_placeholder: true,
-    }
-}
-
-#[test]
-fn dashboard_live_template_renders_monitoring_sections() {
-    let html = routes::test_support::render_dashboard_live(sample_dashboard_view())
-        .expect("dashboard live partial renders");
-
-    assert!(html.contains("Execution mix"));
-    assert!(html.contains("Queue mix"));
-    assert!(html.contains("Agent activity"));
-    assert!(html.contains("Live runtime"));
-    assert!(html.contains("feature-dev"));
-}
-
-#[test]
-fn sessions_template_renders_accessible_list_controls() {
-    let detail = sample_session_detail();
-    let detail_html =
-        routes::test_support::render_session_detail(detail.clone()).expect("detail renders");
-    let html = routes::test_support::render_sessions_page(
-        SessionsPageView {
-            mode_label: "Live runtime".into(),
-            mode_tone: "success",
-            live_stream_url: "/sessions/events?session=discord%3Adirect%3Achan-1".into(),
-            batch_export: BatchExportFormView {
-                action_url: "/api/sessions/export".into(),
-                since: "7d".into(),
-                until: "2026-03-10".into(),
-                limit: 25,
-                format_options: vec![
-                    SelectOption {
-                        value: "json".into(),
-                        label: "JSON".into(),
-                        selected: true,
-                    },
-                    SelectOption {
-                        value: "md".into(),
-                        label: "Markdown".into(),
-                        selected: false,
-                    },
-                ],
-                hint: "Provide at least one of since or until.".into(),
-            },
-            sessions: vec![SessionListItem {
-                title: "ops / bridge".into(),
-                subtitle: "feature-dev active · Live runtime".into(),
-                preview: "Investigate the dashboard refresh cycle.".into(),
-                updated_at: "2026-03-10 12:34".into(),
-                badge: "DISCORD".into(),
-                badge_tone: "cyan",
-                page_url: "/sessions?session=ops".into(),
-                active: true,
-            }],
-            selected: detail,
-        },
-        detail_html,
-    )
-    .expect("sessions template renders");
-
-    assert!(html.contains("data-list-shell"));
-    assert!(html.contains("Search sessions"));
-    assert!(html.contains("data-list-item"));
-    assert!(html.contains("data-detail-panel"));
-    assert!(html.contains("Batch export"));
-    assert!(html.contains("action=\"/api/sessions/export\""));
-    assert!(html.contains("name=\"since\""));
-    assert!(html.contains("Export Markdown"));
-    assert!(html.contains("/api/sessions/discord%3Aops%3Abridge/export?format=json"));
-    assert!(!html.contains("hx-get"));
-}
-
-#[test]
-fn session_detail_template_renders_export_actions_with_empty_messages() {
-    let mut detail = sample_session_detail();
-    detail.messages.clear();
-    detail.empty_hint = "This session has no persisted messages yet.".into();
-
-    let html = routes::test_support::render_session_detail(detail).expect("detail renders");
-
-    assert!(html.contains("Export JSON"));
-    assert!(html.contains("Export Markdown"));
-    assert!(html.contains("This session has no persisted messages yet."));
-}
-
-#[test]
-fn queue_detail_template_renders_table_controls() {
-    let html = routes::test_support::render_queue_detail(sample_queue_detail())
-        .expect("queue detail renders");
-
-    assert!(html.contains("data-table-shell"));
-    assert!(html.contains("Search traffic"));
-    assert!(html.contains("data-table-row"));
-    assert!(html.contains("Retries high-low"));
-}
-
-#[test]
-fn schedules_template_renders_form_actions_and_history() {
-    let detail = sample_schedule_detail();
-    let detail_html = routes::test_support::render_schedule_detail(detail.clone())
-        .expect("schedule detail renders");
-    let html = routes::test_support::render_schedules_page(
-        SchedulesPageView {
-            mode_label: "1 active of 1".into(),
-            mode_tone: "success",
-            schedules: vec![ScheduleListItem {
-                title: "nightly-review".into(),
-                subtitle: "feature-dev · default input".into(),
-                preview: "0 0 * * * * · Next 2026-03-11 00:00:00".into(),
-                source_label: "Last 2026-03-10 12:35".into(),
-                status_label: "Enabled".into(),
-                status_tone: "sage",
-                page_url: "/schedules?schedule=nightly-review".into(),
-                active: true,
-            }],
-            selected: detail,
-            new_schedule_url: "/schedules?schedule=__new__".into(),
-        },
-        detail_html,
-    )
-    .expect("schedules template renders");
-
-    assert!(html.contains("Search schedules"));
-    assert!(html.contains("New schedule"));
-    assert!(html.contains("data-list-item"));
-    assert!(html.contains("Pause schedule"));
-    assert!(html.contains("Recent matching runs"));
-}
-
-#[test]
-fn schedule_detail_template_hides_destructive_controls_for_new_schedule() {
-    let html = routes::test_support::render_schedule_detail(sample_new_schedule_detail())
-        .expect("new schedule detail renders");
-
-    assert!(html.contains("Create schedule"));
-    assert!(html.contains("No teams are installed yet."));
-    assert!(html.contains("No matching runs found for this schedule yet."));
-    assert!(!html.contains("Delete schedule"));
-    assert!(!html.contains("Enable schedule"));
-}
-
-#[test]
-fn plugins_template_renders_install_and_lifecycle_controls() {
-    let detail = sample_plugin_detail();
-    let detail_html =
-        routes::test_support::render_plugin_detail(detail.clone()).expect("plugin detail renders");
-    let html = routes::test_support::render_plugins_page(
-        PluginsPageView {
-            mode_label: "1 operational · 0 attention · 0 disabled".into(),
-            mode_tone: "success",
-            filters: vec![
-                PluginFilterItem {
-                    label: "All".into(),
-                    count: 1,
-                    tone: "neutral",
-                    page_url: "/plugins".into(),
-                    active: true,
-                },
-                PluginFilterItem {
-                    label: "Operational".into(),
-                    count: 1,
-                    tone: "success",
-                    page_url: "/plugins?status=operational".into(),
-                    active: false,
-                },
-            ],
-            plugins: vec![PluginListItem {
-                title: "ops-tools".into(),
-                subtitle: "v1.2.3 · OG".into(),
-                preview: "Operational helpers for dashboards and workflows.".into(),
-                status_detail: "Runtime initialized and ready for operators.".into(),
-                search_text: "skill ops-tools/check-health".into(),
-                source_label: "/home/dh/.opengoose/plugins/ops-tools".into(),
-                source_badge: "ops-tools".into(),
-                status_label: "Ready".into(),
-                status_tone: "success",
-                page_url: "/plugins?plugin=ops-tools".into(),
-                active: true,
-            }],
-            selected: detail,
-        },
-        detail_html,
-    )
-    .expect("plugins template renders");
-
-    assert!(html.contains("Search plugins"));
-    assert!(html.contains("Operational · 1"));
-    assert!(html.contains("Install plugin"));
-    assert!(html.contains("Disable plugin"));
-    assert!(html.contains("Runtime verification"));
-    assert!(html.contains("Confirm removal of"));
-}
-
-#[test]
-fn trigger_detail_template_renders_placeholder_create_form() {
-    let html = routes::test_support::render_trigger_detail(sample_trigger_placeholder_detail())
-        .expect("trigger detail renders");
-
-    assert!(html.contains("No triggers configured"));
-    assert!(html.contains("trigger-create-form"));
-    assert!(html.contains("Name, type, and team are required to create a trigger."));
-    assert!(html.contains("Create trigger"));
-}
-
-#[test]
-fn triggers_template_renders_actions_and_forms() {
-    let detail = sample_trigger_detail();
-    let detail_html = routes::test_support::render_trigger_detail(detail.clone())
-        .expect("trigger detail renders");
-    let html = routes::test_support::render_triggers_page(
-        TriggersPageView {
-            mode_label: "1 trigger(s)".into(),
-            mode_tone: "success",
-            triggers: vec![TriggerListItem {
-                title: "on-pr-open".into(),
-                subtitle: "Webhook".into(),
-                team_label: "code-review".into(),
-                status_label: "enabled".into(),
-                status_tone: "success",
-                last_fired: "2026-03-10 12:34".into(),
-                page_url: "/triggers?trigger=on-pr-open".into(),
-                active: true,
-            }],
-            selected: detail,
-        },
-        detail_html,
-    )
-    .expect("triggers template renders");
-
-    assert!(html.contains("Search triggers"));
-    assert!(html.contains("Run test"));
-    assert!(html.contains("Save changes"));
-    assert!(html.contains("Create new trigger"));
-    assert!(html.contains("Delete trigger"));
-}
-
-#[test]
-fn workflows_template_renders_trigger_controls() {
-    let detail = sample_workflow_detail();
-    let detail_html = routes::test_support::render_workflow_detail(detail.clone())
-        .expect("workflow detail renders");
-    let html = routes::test_support::render_workflows_page(
-        WorkflowsPageView {
-            mode_label: "Live registry".into(),
-            mode_tone: "success",
-            workflows: vec![WorkflowListItem {
-                title: "feature-dev".into(),
-                subtitle: "Build and verify product changes with a chained team.".into(),
-                preview: "1/1 enabled · 0 configured · planner · developer".into(),
-                source_label: "Live registry".into(),
-                status_label: "Running".into(),
-                status_tone: "cyan",
-                page_url: "/workflows?workflow=feature-dev".into(),
-                active: true,
-            }],
-            selected: detail,
-        },
-        detail_html,
-    )
-    .expect("workflows template renders");
-
-    assert!(html.contains("Search workflows"));
-    assert!(html.contains("data-bind:workflow-trigger-input"));
-    assert!(html.contains("/workflows/feature-dev/trigger"));
-    assert!(html.contains("Recent runs"));
-}
