+++ crates/opengoose-web/src/tests/template_render/mod.rs
@@ -1,0 +1,4 @@
+mod schedules_plugins_triggers;
+mod sessions_queue;
+mod support;
+mod workflows_dashboard;
